Changing API

+++ main.py
@@ -1,6 +1,7 @@
 from fastapi import FastAPI, UploadFile, File, Request, Form
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
+from fastapi.staticfiles import StaticFiles
 from song_clustering import ClusterMaker
 app = FastAPI()
 templates = Jinja2Templates(directory="templates")
@@ -22,9 +23,17 @@ model_metadata = {
     "version": "1.0",
     "training_date": "2024-05-19",
     "dataset_used": "Covers80",
-    "dataset_url": "http://labrosa.ee.columbia.edu/projects/coversongs/covers80/"  # Adjust URL as appropriate
-}
+    "dataset_url": "http://labrosa.ee.columbia.edu/projects/coversongs/covers80/",
+    "generated_image_path": "/static/picture.jpg",
 
+    "dataset_description": """
+        As described on the main coversongs page, we have been researching automatic detection 
+        of "cover songs" i.e., alternative performances of the same basic musical piece by 
+        different artists, typically with large stylistic and/or harmonic changes. The 
+        covers80 dataset is a collection of 80 songs, each performed by 2 artists.
+    """
+}
+app.mount("/static", StaticFiles(directory="static"), name="static")
 # Endpoint for the main page
 @app.get("/", response_class=HTMLResponse)
 def read_root(request: Request):
@@ -44,16 +53,25 @@ def predict_form(request: Request):
 @app.post("/predict", response_class=HTMLResponse)
 async def predict(request: Request, audio_file: UploadFile = File(...)):
     contents = await audio_file.read()
-    temp_file_path = "temp_audio_file.mp3"  # Ensure this is the intended format and handling
+    temp_file_path = "temp_audio_file.mp3"  # Save the uploaded file temporarily
     with open(temp_file_path, "wb") as f:
         f.write(contents)
-    prediction = matcher(temp_file_path)  # Assuming matcher now takes raw audio data and sample rate
-
+    
+    # Ensure matcher returns a list of class names sorted by similarity
+    similar_classes = matcher(temp_file_path)  # This should now return a list of names
+    
+    # Clean up temporary files
     import os
     os.remove(temp_file_path)
-    os.remove(temp_file_path.replace("mp3", "npy"))
+    if os.path.exists(temp_file_path.replace("mp3", "npy")):
+        os.remove(temp_file_path.replace("mp3", "npy"))
 
-    return templates.TemplateResponse("prediction_result.html", {"request": request, "prediction": prediction})
+    # Pass the list of similar classes to the template
+    return templates.TemplateResponse("prediction_result.html", {
+        "request": request,
+        "similar_classes": similar_classes  # Make sure to change 'prediction' to 'similar_classes'
+    })
+
 
 # Additional setup for running the application
 if __name__ == "__main__":
